+++ app/guid-file/route.ts
@@ -1,20 +1,21 @@
 import { action } from '@ember-decorators/object';
 import { service } from '@ember-decorators/service';
-import Route from '@ember/routing/route';
 import HeadTagsService from 'ember-cli-meta-tags/services/head-tags';
 import { task } from 'ember-concurrency';
+import moment from 'moment';
+
 import GuidFileController from 'ember-osf-web/guid-file/controller';
 import File from 'ember-osf-web/models/file';
 import Institution from 'ember-osf-web/models/institution';
 import User from 'ember-osf-web/models/user';
+import ResolvedGuidRoute from 'ember-osf-web/resolve-guid/resolved-guid-route';
 import Analytics, { analyticPrivacy } from 'ember-osf-web/services/analytics';
 import CurrentUser from 'ember-osf-web/services/current-user';
 import MetaTags, { HeadTagDef } from 'ember-osf-web/services/meta-tags';
 import Ready from 'ember-osf-web/services/ready';
 import loadAll from 'ember-osf-web/utils/load-relationship';
-import moment from 'moment';
 
-export default class GuidFile extends Route {
+export default class GuidFile extends ResolvedGuidRoute {
     @service analytics!: Analytics;
     @service currentUser!: CurrentUser;
     @service('head-tags') headTagsService!: HeadTagsService;
@@ -42,7 +43,7 @@ export default class GuidFile extends Route {
 
     async model(this: GuidFile, params: { file_guid: string }) {
         try {
-            const file: File = await this.store.findRecord('file', params.file_guid);
+            const file: File = await this.get('resolveGuid').perform(params.file_guid, 'file');
             const fileId = file.get('id');
             const fileUser: User = await file.get('user');
             const user: User = await fileUser.reload();