+++ app/guid-preprint/route.ts
@@ -3,7 +3,7 @@ import ResolvedGuidRoute from 'ember-osf-web/resolve-guid/resolved-guid-route';
 export default class GuidPreprint extends ResolvedGuidRoute {
     model(this: GuidPreprint, params: { preprint_guid: string }) {
         return {
-            taskInstance: this.get('loadModel').perform('preprint', params.preprint_guid),
+            taskInstance: this.get('resolveGuid').perform(params.preprint_guid, 'preprint'),
             preprintId: params.preprint_guid,
         };
     }